+++ src/python/pants/engine/legacy/BUILD
@@ -78,11 +78,3 @@ python_library(
     'src/python/pants/scm:change_calculator',
   ]
 )
-
-python_library(
-  name='spec_parser',
-  sources=['spec_parser.py'],
-  dependencies=[
-    'src/python/pants/base:specs',
-  ]
-)
